Refactor feaLib builder classes into otllib.lookupBuilders
